separate server binding to port from other app config
this seems a bit crude but allows running the tests and having them take
care of setting up the port to use
