+++ test/api.test.ts
@@ -1,7 +1,8 @@
 import {} from "jest";
 import * as supertest from "supertest";
 
-const request = supertest("http://localhost:3000");
+const app = require("../src/app");
+const request = supertest(app);
 
 describe("GET /api", () => {
   it("should return 200 OK", () => {
